fix GSettings pkg-config dependency
We're using the C++ interface of GSettings, so we should depend on
giomm instead of gio.

Since the compile and link flags are different between gio and giomm, it
may seem strange that building paprefs worked at all, but since gtkmm
depends on giomm, the correct flags got added via gtkmm.