rename "module" to "module-group" in GSettings schemas and paths
This change was done in PulseAudio, because a "module" object actually
referred up to 10 module instances, so talking about a module group is
less confusing.

Also rename the PA_GSETTINGS_PATH_MODULES define to simpler
MODULE_GROUPS_PATH and actually use it (previously the define was not
used anywhere), and add MODULE_GROUP_SCHEMA to be used instead of
repeating the same string literal over and over.